requirements.txt
# Core dependencies
python-telegram-bot==21.0
google-genai==0.2.3
pydantic-settings==2.7.1

# Utilities
python-dotenv==1.0.1
